tests/wpt/web-platform-tests/css/css-typed-om/stylevalue-serialization/cssUrlImageValue.tentative.html

<!doctype html>
<meta charset="utf-8">
<title>IDL-constructed CSSURLImageValue serialization tests</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#urlimagevalue-serialization">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/testhelper.js"></script>
<script>
'use strict';

test(() => {
  assert_equals(new CSSURLImageValue('http://foo.bar').toString(), 'url("http://foo.bar")');
}, 'CSSUrlImageValue constructed from IDL serializes correctly');

</script>
